tests: refactor test functions
